Begin phasing out the old tessellator.
