+++ cli/src/commands.rs
@@ -7,7 +7,6 @@ use lyon::algorithms::hatching::{HatchingOptions, DotOptions};
 pub enum Tessellator {
     Default,
     Tess2,
-    Experimental,
 }
 
 #[derive(Clone, Debug)]
