+++ cli/src/tessellate/mod.rs
@@ -43,9 +43,6 @@ pub fn tessellate_path(cmd: TessellateCmd) -> Result<VertexBuffers<Point, u16>, 
                     &mut BuffersBuilder::new(&mut buffers, Identity)
                 ).is_ok()
             }
-            Tessellator::Experimental => {
-                false
-            }
         };
